Moving tasks to celery_tasks
