+++ tests/test_registrations/test_views.py
@@ -63,7 +63,7 @@ class TestRegistrationViews(RegistrationsTestBase):
         res = self.app.get(url, auth=None)
         assert_equal(res.status_code, http.OK)
 
-    @mock.patch('framework.tasks.handlers.enqueue_task', mock.Mock())
+    @mock.patch('framework.celery_tasks.handlers.enqueue_task', mock.Mock())
     def test_register_template_page_backwards_comptability(self):
         # Historically metaschema's were referenced by a slugified version
         # of their name.
@@ -158,7 +158,7 @@ class TestDraftRegistrationViews(RegistrationsTestBase):
         assert_equal(res.status_code, http.ACCEPTED)
         assert_equal(mock_register_draft.call_args[0][0]._id, self.draft._id)
 
-    @mock.patch('framework.tasks.handlers.enqueue_task')
+    @mock.patch('framework.celery_tasks.handlers.enqueue_task')
     def test_register_template_make_public_creates_pending_registration(self, mock_enquque):
         url = self.node.api_url_for('register_draft_registration', draft_id=self.draft._id)
         res = self.app.post_json(url, self.immediate_payload, auth=self.user.auth)
@@ -171,7 +171,7 @@ class TestDraftRegistrationViews(RegistrationsTestBase):
         # The registration created is public
         assert_true(reg.is_pending_registration)
 
-    @mock.patch('framework.tasks.handlers.enqueue_task')
+    @mock.patch('framework.celery_tasks.handlers.enqueue_task')
     def test_register_template_make_public_makes_children_pending_registration(self, mock_enqueue):
         comp1 = NodeFactory(parent=self.node)
         NodeFactory(parent=comp1)
@@ -187,7 +187,7 @@ class TestDraftRegistrationViews(RegistrationsTestBase):
             assert_true(node.is_registration)
             assert_true(node.is_pending_registration)
 
-    @mock.patch('framework.tasks.handlers.enqueue_task')
+    @mock.patch('framework.celery_tasks.handlers.enqueue_task')
     def test_register_draft_registration_with_embargo_creates_embargo(self, mock_enquque):
         url = self.node.api_url_for('register_draft_registration', draft_id=self.draft._id)
         end_date = dt.datetime.utcnow() + dt.timedelta(days=3)
@@ -210,7 +210,7 @@ class TestDraftRegistrationViews(RegistrationsTestBase):
         assert_true(reg.is_pending_embargo)
         assert_false(reg.embargo_end_date)
 
-    @mock.patch('framework.tasks.handlers.enqueue_task')
+    @mock.patch('framework.celery_tasks.handlers.enqueue_task')
     def test_register_draft_registration_with_embargo_adds_to_parent_project_logs(self, mock_enquque):
         initial_project_logs = len(self.node.logs)
         res = self.app.post_json(
@@ -224,7 +224,7 @@ class TestDraftRegistrationViews(RegistrationsTestBase):
         # Logs: Created, registered, embargo initiated
         assert_equal(len(self.node.logs), initial_project_logs + 1)
 
-    @mock.patch('framework.tasks.handlers.enqueue_task')
+    @mock.patch('framework.celery_tasks.handlers.enqueue_task')
     def test_register_draft_registration_with_embargo_is_not_public(self, mock_enqueue):
         res = self.app.post_json(
             self.node.api_url_for('register_draft_registration', draft_id=self.draft._id),
@@ -241,7 +241,7 @@ class TestDraftRegistrationViews(RegistrationsTestBase):
         assert_true(registration.is_pending_embargo)
         assert_is_not_none(registration.embargo)
 
-    @mock.patch('framework.tasks.handlers.enqueue_task')
+    @mock.patch('framework.celery_tasks.handlers.enqueue_task')
     def test_register_draft_registration_invalid_embargo_end_date_raises_HTTPError(self, mock_enqueue):
         res = self.app.post_json(
             self.node.api_url_for('register_draft_registration', draft_id=self.draft._id),
