+++ website/notifications/tasks.py
@@ -4,7 +4,7 @@ Tasks for making even transactional emails consolidated.
 from bson.code import Code
 from modularodm import Q
 
-from framework.tasks import app as celery_app
+from framework.celery_tasks import app as celery_app
 from framework.mongo import database as db
 from framework.auth.core import User
 from framework.sentry import log_exception
